// src/app/interfaces/global-market.ts
export interface GlobalMarket {
  currentPrice: number;
  ath: number;
  athPercentage: number;
  atl: number;
  atlPercentage: number;
  marketCap: number;
  totalVolume: number;
  high24: number;
  low24: number;
  priceChange24: number;
  priceChange24Percentage: number;
  totalSupply: number;
  circulatingSupply: number;
}